# src/books/build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"BookID": {
				"type": "string",
				"description": "A unique identifier for a particular book"
			},
			"Book": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/BookID"
					},
					"name": {
						"type": "string"
					},
					"author": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"image": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"author",
					"description",
					"price",
					"image"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Filter": {
				"properties": {
					"from": {
						"type": "number",
						"format": "double"
					},
					"to": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"author": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "books",
		"version": "1.0.0",
		"description": "This is a template repository for the first assignment in BVD 103 at McMaster Continuing Education.",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/books/list": {
			"post": {
				"operationId": "ListBooks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Book"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/Filter"
								},
								"type": "array"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetBooks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Book"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "priceFrom",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "priceTo",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/books": {
			"post": {
				"operationId": "CreateOrUpdateBook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"id": {
											"$ref": "#/components/schemas/BookID"
										}
									},
									"required": [
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Book"
							}
						}
					}
				}
			}
		},
		"/books/{book}": {
			"get": {
				"operationId": "GetBook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Book"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "book",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/BookID"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}